Divide From to smaller components
